Refactoring
- Added filtering to resoutce type item summary.
- Corrected namespace.
