+++ app/Models/ResourceTypeItem/Summary/SimpleItem.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace App\Models\ResourceTypeItem\Summary;
 
-use App\Interfaces\ResourceTypeItemType\ISummaryModel;
+use App\Interfaces\ResourceTypeItem\ISummaryModel;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Query\Builder as QueryBuilder;
